CRUD modelos
haciendo las relaciones de modelos con otros modelos y materia prima

+++ app/Http/Controllers/DocumentoController.php
@@ -138,6 +138,9 @@ class DocumentoController extends Controller
     public function destroy(Request $request)
     {
         $documento = Documento::find($request->boton_delete);
+        if( !$documento->proveedores->isEmpty()||( $documento->user!=null)){
+            return redirect()->back()->withErrors(['message2'=>'No se puede eliminar el documento porque esta relacionado']);
+        }
         $documento->delete();
         return redirect()->back();
     }
